// src/screens/tools/effects/Glassmorphism/types.ts
export type TGlassmorphismConfig = {
  backdropBlur: number;
  backgroundColor: string;
  backgroundOpacity: number;
  borderWidth: number;
  borderColor: string;
  borderOpacity: number;
  borderRadius: number;
  boxShadow: boolean;
  elementWidth: number;
  elementHeight: number;
  sceneBackgroundColor: string;
};